[FEATURE] Enable/disable cascading deletion of child elements
Deleting a record will also delete all child records,
which are configured as type inline (field, list) within.
This patch adds a behaviour configuration option "enableCascadingDelete"
which is by default treated as true. If set to FALSE or '' or 0
deletion of inline child records will be disabled.

Resolves: #50391
Releases: 6.2

+++ typo3/sysext/core/Classes/DataHandling/DataHandler.php
@@ -4524,10 +4524,18 @@ class DataHandler {
 					$dbAnalysis = GeneralUtility::makeInstance('TYPO3\\CMS\\Core\\Database\\RelationHandler');
 					$dbAnalysis->start($value, $conf['foreign_table'], '', $uid, $table, $conf);
 					$dbAnalysis->undeleteRecord = TRUE;
+
+					$enableCascadingDelete = TRUE;
+					if (isset($conf['behaviour']['enableCascadingDelete']) && !$conf['behaviour']['enableCascadingDelete']) {
+						$enableCascadingDelete = FALSE;
+					}
+
 					// Walk through the items and remove them
 					foreach ($dbAnalysis->itemArray as $v) {
 						if (!$undeleteRecord) {
-							$this->deleteAction($v['table'], $v['id']);
+							if ($enableCascadingDelete) {
+								$this->deleteAction($v['table'], $v['id']);
+							}
 						} else {
 							$this->undeleteRecord($v['table'], $v['id']);
 						}
